requirements.txt
Flask>=2.1.1
tensorflow>=2.12.0
Pillow>=8.4.0
werkzeug>=2.1.1
gunicorn